+++ src/main/java/com/matzip/thread/users/application/port/in/SignUpRequest.java
@@ -1,25 +1,25 @@
 package com.matzip.thread.users.application.port.in;
 
-import com.matzip.thread.users.application.port.out.UserEntity;
-import com.matzip.thread.users.domain.Membership;
+import com.matzip.thread.users.domain.Role;
 import com.matzip.thread.users.domain.User;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 public class UserCreateRequest {
     private final String username;
     private final String password;
-    private final Membership membership;
+    private final Role role;
 
-    public UserCreateRequest(String username, String password, Membership membership) {
+    public UserCreateRequest(String username, String password, Role role) {
         this.username = username;
         this.password = password;
-        this.membership = membership;
+        this.role = role;
     }
 
     public User toDomainEntity() {
         return new User(
                 username,
                 password,
-                membership
+                role
         );
     }
 
@@ -27,7 +27,11 @@ public class UserCreateRequest {
         return new UserCreateRequest(
                 user.getUsername(),
                 user.getPassword(),
-                user.getMembership()
+                user.getRole()
         );
+    }
+
+    public UserCreateRequest passwordEncode(PasswordEncoder passwordEncoder) {
+        return new UserCreateRequest(username, passwordEncoder.encode(password), role);
     }
 }
